+++ src/features/assetsDecorator.ts
@@ -62,6 +62,7 @@ export function activate(context: vscode.ExtensionContext) {
       while ((match = regex.exec(text))) {
         const startPos = activeEditor.document.positionAt(match.index);
         const endPos = activeEditor.document.positionAt(match.index + match[0].length);
+
         const text = onMatch(match);
         if (!text) {
           continue;
@@ -83,13 +84,27 @@ export function activate(context: vscode.ExtensionContext) {
       activeEditor.setDecorations(type, guids);
     };
 
-    traverse(activeEditor, 'editorCodeLens.foreground', /<(\w+)>(\d+)<\/\1>/g, (match) => {
-      return decorationText(match[1], match[2]);
+    let withinStandard = false;
+    traverse(activeEditor, 'editorCodeLens.foreground', /(<\/?Standard>)|<(\w+)>(\d+)<\/\2>/g, (match) => {
+      if (match[1] === '<Standard>') {
+        withinStandard = true;
+      }
+      else if (match[1] === '</Standard>') {
+        withinStandard = false;
+      }
+      else if (!withinStandard || (match[2] !== 'GUID' && match[2] !== 'Name')) {
+        return decorationText(match[2], match[3]);
+      }
+
+      return '';
     }, guidDecorationType);
 
-    traverse(activeEditor, 'terminal.ansiGreen' /*'button.background'*/, /<Asset>/g, (match) => {
-      const startPos = activeEditor!.document.positionAt(match.index);
-      let standard = activeEditor!.document.getText(new vscode.Range(startPos.line, startPos.character, startPos.line + 20, 0));
+    traverse(activeEditor, 'editorCodeLens.foreground', /<Asset>/g, (match) => {
+      const doc = activeEditor?.document;
+      if (!doc) return '';
+
+      const startPos = doc.positionAt(match.index);
+      let standard = doc.getText(new vscode.Range(startPos.line, startPos.character, startPos.line + 20, 0));
       const endPos = standard.indexOf('</Asset>');
       if (endPos >= 0) {
         standard = standard.substring(0, endPos);